requirements.txt
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2==2.9.6
PyJWT==2.7.0
pytest==7.4.0
pytz==2023.3
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
